requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
transformers
ollama-python
python-dotenv
